fix(examples): align chat example in Java
# Conflicts:
#	examples/simple-examples/build.gradle.kts

# Conflicts:
#	examples/simple-examples/build.gradle.kts
